processing/inferencing/requirements.txt
pandas==1.1.3
numpy==1.19.2
snowflake-connector-python[pandas]==2.6.1
boto3==1.16.32
pandas-profiling==3.0.0
jsonschema==3.2.0
PyYAML==5.4.1
sagemaker==2.59.2